refactor: extract cluster allocator into dev/alloc.rs
Move HostCluster and the refcount-based cluster allocator (grow_reftable,
free_clusters, refblock cache helpers, allocate_clusters and friends)
plus the three allocator unit tests into a new dev/alloc.rs submodule as
an impl extension of Qcow2Dev. Methods and HostCluster called across the
new module boundary are widened from private to pub(crate); this is the
mechanical cost of splitting the single Qcow2Dev impl block. No logic
changes.
